refactor(shared core): refactor shared core library
replaced 'data-access' directory with 'services';
shared-core library should be decomposed further;
