+++ apps/nx-ng-starter/src/app/components/app/app.component.ts
@@ -1,7 +1,7 @@
 import { ChangeDetectionStrategy, Component, Inject, ViewChild } from '@angular/core';
 import { MatSidenav } from '@angular/material/sidenav';
 import { Title } from '@angular/platform-browser';
-import { APP_ENV, WebEnvironment } from '@nx-ng-starter/shared-core/data-access';
+import { APP_ENV, WebEnvironment } from '@nx-ng-starter/shared-core/services';
 import { SidebarUiService } from '@nx-ng-starter/shared-store/state';
 
 /**
