+++ apps/nx-ng-starter/src/environments/environment.prod.ts
@@ -1,4 +1,4 @@
-import { WebEnvironment } from '@nx-ng-starter/shared-core/data-access';
+import { WebEnvironment } from '@nx-ng-starter/shared-core/services';
 
 /**
  * Production environment variables.
