+++ libs/mocks-core/src/lib/data-access/http-handlers/http-handlers-provider.mock.ts
@@ -7,7 +7,7 @@ import {
   ToasterService,
   toasterServiceProvider,
   WebEnvironment,
-} from '@nx-ng-starter/shared-core/data-access';
+} from '@nx-ng-starter/shared-core/services';
 import {
   HttpProgressService,
   httpProgressServiceProvider,
