+++ libs/shared-core/src/lib/services/interfaces/environment/environment.interface.ts
