+++ libs/shared-core/src/lib/services/interfaces/index.ts
