+++ libs/shared-core/src/lib/services/interfaces/ui/ui-dictionary.interface.ts
