+++ libs/shared-core/src/lib/services/interfaces/ui/ui-translations.interface.ts
