+++ libs/shared-core/src/lib/services/interfaces/user/user.interface.ts
