+++ libs/shared-core/src/lib/services/shared-core-services.module.ts
