+++ libs/shared-store/src/lib/state/http-api/http-api.service.spec.ts
@@ -6,7 +6,7 @@ import {
   httpHandlersProviders,
   newTestBedMetadata,
 } from '@nx-ng-starter/mocks-core';
-import { HttpHandlersService, ToasterService } from '@nx-ng-starter/shared-core/data-access';
+import { HttpHandlersService, ToasterService } from '@nx-ng-starter/shared-core/services';
 import { AppTranslateModule, CustomMaterialModule } from '@nx-ng-starter/shared-ui';
 import { Apollo, ApolloModule } from 'apollo-angular';
 import { HttpLinkModule } from 'apollo-angular-link-http';
