+++ libs/shared-store/src/lib/state/http-api/http-api.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable, Provider } from '@angular/core';
 import { Store } from '@ngxs/store';
-import { HttpHandlersService } from '@nx-ng-starter/shared-core/data-access';
+import { HttpHandlersService } from '@nx-ng-starter/shared-core/services';
 import { Message } from '@nx-ng-starter/shared-util';
 import { tap } from 'rxjs/operators';
