+++ tsconfig.base.json
@@ -35,7 +35,7 @@
       "@nx-ng-starter/mocks-core/util": ["libs/mocks-core/src/lib/util/index.ts"],
       "@nx-ng-starter/proto": ["libs/proto/src/index.ts"],
       "@nx-ng-starter/shared-core": ["libs/shared-core/src/index.ts"],
-      "@nx-ng-starter/shared-core/data-access": ["libs/shared-core/src/lib/data-access/index.ts"],
+      "@nx-ng-starter/shared-core/services": ["libs/shared-core/src/lib/services/index.ts"],
       "@nx-ng-starter/shared-store": ["libs/shared-store/src/index.ts"],
       "@nx-ng-starter/shared-store/*": ["libs/shared-store/src/lib/*"],
       "@nx-ng-starter/shared-ui": ["libs/shared-ui/src/index.ts"],
